requirements.txt
Django==1.6.4
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.5
django-storages==1.1.8
django-toolbelt==0.0.1
djangorestframework==2.3.14
gunicorn==18.0
psycopg2==2.5.2
pystache==0.5.3
static==1.0.2
wsgiref==0.1.2
